Move QueryCriteria and validation into GeoQuery.ts.
This provides better modularization, making it easier to import only the needed types.
